requirements.txt
python-telegram-bot
openai
requests
python-dotenv
google-generativeai
httpx
aiogram
nest_asyncio